some bug is fixed

+++ src/main/java/com/finalProject/Project/entity/dto/OrderDto.java
@@ -24,7 +24,7 @@ public class OrderDto {
     private String expertName;
     private Integer offerId;
     private Integer subServiceId;
-    private Integer rate;
+    private Float rate;
     private String commentText;
     private Integer  customerId;
     private Integer amount;
